+++ src/application/people.ts
@@ -1,7 +1,16 @@
-import { PathFactoryFunction } from '../pco-api-client/types.js';
+import { PathFactoryFunction, pathFactoryProvider } from '../pco-api-client/types.js';
 import { ApplicationConfig } from './application.js';
 
-const paths = new Map<string, PathFactoryFunction>([]);
+const paths = new Map<string, PathFactoryFunction>([
+  // Root
+  ['people', pathFactoryProvider('get', '')],
+
+  // Emails
+  ['readEmails', pathFactoryProvider('get', 'emails')],
+
+  // Phone Numbers
+  ['readPhoneNumbers', pathFactoryProvider('get', 'phone_numbers')],
+]);
 
 export const PeopleConfigDefault: ApplicationConfig = {
   apiVersion: 'v2',